rename users hook to user
